+++ lotti/lib/pages/settings/measurables/measurables_page.dart
@@ -1,13 +1,12 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/settings/measurables/measurable_type_card.dart';
+import 'package:lotti/routes/router.gr.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/utils/file_utils.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
-
-import 'measurable_detail_page.dart';
 
 const double iconSize = 24.0;
 
@@ -91,27 +90,7 @@ class _MeasurablesPageState extends State<MeasurablesPage> {
                   child: const Icon(MdiIcons.plus, size: 32),
                   backgroundColor: AppColors.entryBgColor,
                   onPressed: () {
-                    Navigator.of(context).push(
-                      MaterialPageRoute(
-                        builder: (BuildContext context) {
-                          DateTime now = DateTime.now();
-                          return DetailRoute(
-                            item: MeasurableDataType(
-                              id: uuid.v1(),
-                              name: '',
-                              displayName: '',
-                              version: 0,
-                              createdAt: now,
-                              updatedAt: now,
-                              unitName: '',
-                              description: '',
-                              vectorClock: null,
-                            ),
-                            index: -1,
-                          );
-                        },
-                      ),
-                    );
+                    context.router.push(const CreateMeasurableRoute());
                   },
                 ),
               ),
