+++ lotti/lib/routes/settings_routes.dart
@@ -7,6 +7,8 @@ import 'package:lotti/pages/settings/flags_page.dart';
 import 'package:lotti/pages/settings/health_import_page.dart';
 import 'package:lotti/pages/settings/insights_page.dart';
 import 'package:lotti/pages/settings/maintenance_page.dart';
+import 'package:lotti/pages/settings/measurables/measurable_create_page.dart';
+import 'package:lotti/pages/settings/measurables/measurable_details_page.dart';
 import 'package:lotti/pages/settings/measurables/measurables_page.dart';
 import 'package:lotti/pages/settings/outbox_monitor.dart';
 import 'package:lotti/pages/settings/settings_page.dart';
@@ -59,6 +61,14 @@ const AutoRoute settingsRoutes = AutoRoute(
     AutoRoute(
       path: 'measurables',
       page: MeasurablesPage,
+    ),
+    AutoRoute(
+      path: 'measurables/:measurableId',
+      page: EditMeasurablePage,
+    ),
+    AutoRoute(
+      path: 'measurables/create',
+      page: CreateMeasurablePage,
     ),
     AutoRoute(
       path: 'outbox_monitor',
